pyproject.toml
[tool.poetry]
name = "skyshop"
version = "0.1.0"
description = ""
authors = ["Skillova <skillova@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
black = "^24.4.2"
pillow = "^10.3.0"
ipython = "^8.24.0"
pytils = "^0.4.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}


[tool.poetry.group.dev.dependencies]
setuptools = "^70.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
